RUMM-180 Extract the SystemInfo impl to core
